// src/types.d.ts
export interface City {
  id: number;
  name: string;
  country: string;
  latitude: number;
  longitude: number;
}

export interface Welcome {
    latitude:              number;
    longitude:             number;
    generationtime_ms:     number;
    utc_offset_seconds:    number;
    timezone:              string;
    timezone_abbreviation: string;
    elevation:             number;
    current_units:         CurrentUnits;
    current:               Current;
    daily_units:           DailyUnits;
    daily:                 Daily;
    hourly_units:          HourlyUnits;
    hourly:                Hourly;
}

export interface Hourly {
    time:                  string[];
    temperature_2m:        number[];
    weather_code:          number[];
}

export interface HourlyUnits {
    time:                  string;
    temperature_2m:        string;
    weather_code:          string;
}

export interface Daily {
    time:                  string[];
    weather_code:          number[];
    temperature_2m_max:    number[];
    temperature_2m_min:    number[];
}

export interface DailyUnits {
    time:                  string;
    weather_code:          string;
    temperature_2m_max:    string;
    temperature_2m_min:    string;
}

export interface Current {
    time:                  string;
    interval:              number;
    temperature_2m:        number;
    relative_humidity_2m:  number;
    apparent_temperature:  number;
    precipitation:         number;
    wind_speed_10m:        number;
    weather_code:          number;
    is_day:                number;
}

export interface CurrentUnits {
    time:                  string;
    interval:              string;
    temperature_2m:        string;
    relative_humidity_2m:  string;
    apparent_temperature:  string;
    precipitation:         string;
    wind_speed_10m:        string;
    weather_code:          string;
    is_day:                string;
}